Reuse compress buffer (#60)

+++ src/page/page_dict/fixed_len_binary.rs
@@ -44,8 +44,8 @@ fn read_plain(bytes: &[u8], size: usize, length: usize) -> Vec<u8> {
     bytes[..size * length].to_vec()
 }
 
-pub fn read(buf: &[u8], size: i32, num_values: u32) -> Result<Arc<dyn DictPage>> {
-    let values = read_plain(buf, size as usize, num_values as usize);
+pub fn read(buf: &[u8], size: i32, num_values: usize) -> Result<Arc<dyn DictPage>> {
+    let values = read_plain(buf, size as usize, num_values);
     Ok(Arc::new(FixedLenByteArrayPageDict::new(
         values,
         PhysicalType::FixedLenByteArray(size),
